Moved updatePassword to SettingsController and improved it
