+++ app/Http/Controllers/Frontend/SettingsController.php
@@ -1,0 +1,59 @@
+<?php
+
+namespace App\Http\Controllers\Frontend;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
+
+class SettingsController extends Controller
+{
+
+    public function updateMainSettings(Request $request): RedirectResponse {
+        $validated = $request->validate([
+                                            'username' => ['required', 'string', 'max:25', 'regex:/^[a-zA-Z0-9_]*$/'],
+                                            'name'     => ['required', 'string', 'max:50'],
+                                            'email'    => ['required', 'string', 'email', 'max:255'],
+                                            'avatar'   => 'image'
+                                        ]);
+
+        if (auth()->user()->username != $request->username) {
+            $request->validate(['username' => ['unique:users']]);
+        }
+        if (auth()->user()->email != $request->email) {
+            $request->validate(['email' => ['unique:users']]);
+            auth()->user()->update(['email_verified_at' => null]);
+        }
+        auth()->user()->update([
+                                   'email'           => $validated['email'],
+                                   'username'        => $validated['username'],
+                                   'name'            => $validated['name'],
+                                   'always_dbl'      => $request->always_dbl == "on",
+                                   'private_profile' => $request->private_profile == "on",
+                               ]);
+
+        if (!auth()->user()->hasVerifiedEmail()) {
+            auth()->user()->sendEmailVerificationNotification();
+        }
+
+        return back();
+    }
+
+    public function updatePassword(Request $request): RedirectResponse {
+        $validated = $request->validate([
+                                            'currentPassword' => ['required'],
+                                            'password'        => ['required', 'string', 'min:8', 'confirmed']
+                                        ]);
+
+        if (!Hash::check($validated['currentPassword'], auth()->user()->password)) {
+            return back()->withErrors(__('controller.user.password-wrong'));
+        }
+
+        auth()->user()->update([
+                                   'password' => Hash::make($validated['password'])
+                               ]);
+
+        return back()->with('info', __('controller.user.password-changed-ok'));
+    }
+}
